+++ spec/generators/fixtures/basic_rails_app/config/application.rb
@@ -3,6 +3,12 @@
 require File.expand_path("../boot", __FILE__)
 
 require "rails"
+
+# We need Ruby Next since this code is loaded in generator tests in
+# a separate process
+require "ruby-next/language/runtime"
+RubyNext::Language.watch_dirs << File.expand_path("../../../../lib", __FILE__)
+
 require "action_controller/railtie"
 require "anyway_config"